modularization: move file-related structs to module io
This build upon the good work Ivan has done. It is nicer to find all
file-related structs at a single module.
